fix(constants): rename typo filename
